extract the Microsoft BusyStatus
this makes the BusyStatus custom attribute presented by Microsoft
available as property 'event.ms_busystatus'.

Events that don't have this property in it's source will always get the
status 'BUSY' that most events in a Outlook-Driven environment would
have.

The possible values found in the Microsoft documentation are:
 - 'FREE' - User marked this event as not affecting his availability to others
 - 'TENTATIVE' - User is not sure if he/she will be in attendence
 - 'BUSY' - User has decided to attend
 - 'OOF' - User is "not in the office" during this event

The use case for myself is to only include BUSY events on my Mirror, as
all other types are either not fix yet or don't actually affect my
schedule.