add a way to deduplicate events
this seems to be the best way in my understanding, allowing the user to
decide on which event properties a deduplication should be decided on.

Of cause using this has a gotcha: the calendarName property is being
modified. This is purely driven by my usecase. But this could be made
configurable as well (thinking about config.deduplicationMergeProps).