Re-try parsing even if there was an error with the request for ical data
Previously if there was an intermittent connectivity failure when fetching
the ical data (WiFI outage for example) the calendar(s) affected by the
problem would stop refreshing.

Obviously this means that it's possible a bad calendar will be tried indefintely,
so a better approach might be to use some sort of exponential back-off between
retries before eventually giving up after a configurable time window... but this
change is much simpler!