config: protect against user errors
When a user accidentially enters a string like '10*60*1000' into a field
which we pass straight through to setTimeout() it will be cast to an
integer, but the integer will be (in this example) '10', which is way
too short for any reasonable hardware, causing e.g. rendering glitches.

In order to fix this (in most cases) manually execute the intended
javascript code, which evalues the (typically a math) expression and
yields the behaviour the user expects.

Fixes #74